working lightgbm with flats

+++ gahaco/models/train_model.py
@@ -38,8 +38,6 @@ from gahaco.features.correlation import select_uncorrelated_features
 # Flags 
 # -----------------------------------------------------------------------------
 flags.DEFINE_string('model', 'rnf', 'model to run') # name ,default, help
-flags.DEFINE_string('models_dir', '/cosma6/data/dp004/dc-cues1/tree_models/',
-							'dir to save models for error analysis')
 flags.DEFINE_integer('np', 1, 'Number of processes to run') 
 flags.DEFINE_integer('n_splits', 4, 'Number of folds for cross-validation') 
 flags.DEFINE_boolean('upload', False, 'upload model to comet ml, otherwise save in temporary folder') 
@@ -60,10 +58,10 @@ def main(argv):
 			api_key="VNQSdbR1pw33EkuHbUsGUSZWr", project_name="general", workspace="florpi"
 		)
 	else:
-		print('Correctly going offline')
+		print('Going offline, remember to upload if interesting :)')
 		experiment = OfflineExperiment(
 			api_key="VNQSdbR1pw33EkuHbUsGUSZWr", project_name="general", workspace="florpi",
-			offline_directory="/tmp/"
+			offline_directory="/cosma/home/dp004/dc-cues1/GaHaCo/comet/",
 		)
 
 	if FLAGS.logging:
@@ -82,16 +80,26 @@ def main(argv):
 	config = load_config(config_file_path=config_file_path)
 
 	config['model']['parameters']['n_jobs'] = FLAGS.np
+
 	print(f'Using {FLAGS.np} cores to fit models')
-	print(config['model']['parameters'])
+
 	# -------------------------------------------------------------------------
 	# Load and prepare datasets
 	# -------------------------------------------------------------------------
 
 	# Load dataset
 	features, labels = get_data(config["label"])
-	# Select feautures by correlation
-	#features = select_uncorrelated_features(features)
+
+	if "feature_optimization" in config.keys():
+		if config['feature_optimization']['PCA']: 
+			# TODO: Needs to be updated to only take features and return dataframe
+			train_features, test_features = feature_optimization(
+								train, test, config["feature_optimization"], experiment=experiment
+												        )
+
+			feature_names = [f"PCA_{i}" for i in range(train["features"].shape[1])]
+			
+
 	# K-Fold validation
 	metric_module = importlib.import_module(config["metric"]["module"])
 	metric = getattr(metric_module, config["metric"]["method"])
@@ -136,13 +144,6 @@ def main(argv):
 		# Set-up and Run inference model
 		# -------------------------------------------------------------------------
 		trained_model = fit(x_train, y_train, config["model"])
-
-		if FLAGS.models_dir:
-			with open(FLAGS.models_dir+f'{FLAGS.model}_{fold}','wb') as f:
-				    pickle.dump(trained_model, f)
-			x_test.to_hdf(FLAGS.models_dir+f'{FLAGS.model}_{fold}_test_feats.hdf5', key='df', mode='w')
-			y_test.to_hdf(FLAGS.models_dir+f'{FLAGS.model}_{fold}_test_labels.hdf5', key='df', mode='w')
-
 		y_pred = prediction(trained_model, x_test, config["model"])
 		
 		metric_value = metric(y_test, y_pred, **config["metric"]["params"])
@@ -159,10 +160,12 @@ def main(argv):
 		cm = confusion_matrix(y_test, n_hod_galaxies)
 		cm = cm.astype("float") / cm.sum(axis=1)[:, np.newaxis]
 		hod_cms.append(cm)
-		imp = feature_importance.dropcol(trained_model, x_train, y_train,
-					x_test, y_test, metric_value, metric, config['metric']['params'])
-		importances.append(imp)
-		sk_importances.append(trained_model.feature_importances_)
+
+		if config['feature_optimization']['measure_importance']:
+			imp = feature_importance.dropcol(trained_model, x_train, y_train,
+						x_test, y_test, metric_value, metric, config['metric']['params'])
+			importances.append(imp)
+			sk_importances.append(trained_model.feature_importances_)
 
 		test_hydro_pos, test_dmo_pos= load_positions(test_idx)
 		r_c, test_hydro_tpcf = compute_tpcf(test_hydro_pos[y_test>0])
@@ -175,7 +178,6 @@ def main(argv):
 	# Save output's visualizations
 	# -------------------------------------------------------------------------
 
-	print(cms)
 
 	visualize.plot_confusion_matrix(cms,
 									classes = ['Dark', 'Luminous'],
@@ -189,22 +191,24 @@ def main(argv):
 									title='HOD',
 									experiment = experiment,
 									)
-
-	visualize.plot_feature_importance(importances,
-								x_train.columns,
-								title='Drop column',
-								experiment=experiment)
-
-	visualize.plot_feature_importance(sk_importances,
-								x_train.columns,
-								title='Gini impurity',
-								experiment=experiment)
-
 	visualize.plot_tpcfs(
 					r_c, hydro_tpcf, pred_tpcf, hod_tpcf, experiment=experiment
 						)
 
-	experiment.add_tag(f'preprocessing = {config["sampling"]}')
+
+
+	if config['feature_optimization']['measure_importance']:
+		visualize.plot_feature_importance(importances,
+									x_train.columns,
+									title='Drop column',
+									experiment=experiment)
+
+		visualize.plot_feature_importance(sk_importances,
+									x_train.columns,
+									title='Gini impurity',
+									experiment=experiment)
+
+	#experiment.add_tag(f'preprocessing = {config["sampling"]["method]}')
 	experiment.add_tag(f'classifier = {FLAGS.model}')
 
 	print('All good :)')
